% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BIC.R
\name{calculate_BIC}
\alias{calculate_BIC}
\title{calculate_BIC}
\usage{
calculate_BIC(STIE_result, ST_expr)
}
\arguments{
\item{STIE_result}{}

\item{ST_expr}{}
}
\description{
calculate_BIC
}
\examples{

data = mtcars[,c("disp","hp")]
model1 <- lm(mpg ~ disp + hp, data = mtcars)
l = sum(log(dnorm(r,sd=sd(r))))
d = ncol(data) + 2
n = nrow(data)
bic = -2*l + d*log(n)

}
