% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_dist.R
\name{calculate_cell_dist}
\alias{calculate_cell_dist}
\title{calculate_cell_dist}
\usage{
calculate_cell_dist(cells_on_spot, dist_type = c("center", "boundary"),
  dist_cutoff = Inf, axis = c("Minor", "Major"))
}
\arguments{
\item{cells_on_spot}{a data frame representing the cells on spots along with the cellular morphological features}

\item{dist_type}{a boolean value indicating whether the distance calculated between cell centers or cell boundary}

\item{dist_cutoff}{a numeric value indicating the cutoff of maximum distance between two cells}

\item{axis}{a boolean value indicating whether the distance is calculated along major or minor axis}
}
\value{
a matrix of numeric values representing the distance between each cell pair
}
\description{
calculate_cell_dist calculates the distance between cells based on the cell coordinates
}
\references{

}
\seealso{
\code{\link{get_cells_on_spot}};
}
\author{
Shijia Zhu, \email{shijia.zhu@UTsouthwestern.edu}
}
