% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_morphology_probability.R
\name{calculate_morphology_probability}
\alias{calculate_morphology_probability}
\title{calculate_morphology_probability}
\usage{
calculate_morphology_probability(morphology_features, feature_list, mu,
  sigma, scaled = TRUE)
}
\arguments{
\item{morphology_features}{a matrix of numeric values in size of cells*features representing the values of morphology_features for each cell, where the column names are the morphoglical features, and row names are cell ids}

\item{feature_list}{a vector of characters in size of morphology features representing the morphology feature names used in the analysis}

\item{mu}{a vector of numeric values in size of morphology features representing the mean of the normal distribution}

\item{sigma}{a vector of numeric values in size of morphology features representing the standard deviation of the normal distribution}

\item{scaled}{boolean value representing whether the probability is scaled}
}
\value{
a matrix of numeric values in size of cells*features representing the values of morphology feature probability for each cell
}
\description{
calculate_morphology_probability calculate the morphological probability by assuming the normal distribution with parameters mu and sigma
}
\examples{
# 



}
\references{

}
\seealso{
\code{\link{update_morphology_parameter}}; \code{\link{get_cells_on_spot}};
}
\author{
Shijia Zhu, \email{shijia.zhu@UTsouthwestern.edu}
}
