% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RMSE.R
\name{calculate_RMSE}
\alias{calculate_RMSE}
\title{calculate_RMSE}
\usage{
calculate_RMSE(STIE_result, ST_expr, Signature = NULL)
}
\arguments{
\item{STIE_result}{}

\item{ST_expr}{}

\item{Signature}{}
}
\description{
calculate_RMSE
}
\examples{

data = mtcars[,c("disp","hp")]
model1 <- lm(mpg ~ disp + hp, data = mtcars)
l = sum(log(dnorm(r,sd=sd(r))))
d = ncol(data) + 2
n = nrow(data)
bic = -2*l + d*log(n)

}
