% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spot_radius.R
\name{calculate_spot_radius}
\alias{calculate_spot_radius}
\title{calculate_spot_radius}
\usage{
calculate_spot_radius(spot_coordinates, fct)
}
\arguments{
\item{spot_coordinates}{}

\item{fct}{}
}
\description{
calculate_spot_radius
}
