% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errplot.R
\name{errplot}
\alias{errplot}
\title{errplot}
\usage{
errplot(X, barx = c(1:length(X)), xlab = "", ylab = "", line = T,
  ylim = NULL, col = "black")
}
\arguments{
\item{X}{a list of numeric values to plot}

\item{barx}{the coordinates of points in the plot.}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{line}{a boolean value indicating whether to plot lines}

\item{ylim}{range for the y axis}

\item{col}{character values indicating the color of the plot}
}
\description{
errplot plot lines with error bar
}
