% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sig_markers.R
\name{find_sig_markers}
\alias{find_sig_markers}
\title{find_sig_markers}
\usage{
find_sig_markers(STIE_result, ST_expr, transform = c("logcount",
  "inverse.transform"), DEG_pthres = 1e-05)
}
\arguments{
\item{STIE_result}{a list representing the result generated by the STIE() function}

\item{ST_expr}{a matrix of gene expression representing the spot level gene expression, with row being the spot and column being the gene}

\item{transform}{a character value representing the method to transform the count data}

\item{DEG_pthres}{a numeric value representing the pvalue threshold for DEG calling}
}
\value{
a data frame for cell type marker genes with columns of features, diff and pvalue
}
\description{
find_sig_markers estimates the cell type gene markers from spot level gene expression and STIE estimation
}
\seealso{
\code{\link{STIE}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
