% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cellchat.R
\name{get_cellchat}
\alias{get_cellchat}
\title{get_cellchat}
\usage{
get_cellchat(STIE_result, database = c("human", "mouse"),
  db_category = c("Secreted Signaling"), DEG_pthres = 1e-05)
}
\arguments{
\item{STIE_result}{}

\item{database}{}

\item{db_category}{}

\item{DEG_pthres}{}
}
\description{
get_cellchat
}
\examples{



}
