% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_colors.R
\name{get_my_colors}
\alias{get_my_colors}
\title{get_my_colors}
\usage{
get_my_colors(n, mode = 1)
}
\arguments{
\item{n}{an integer value representing the number of colors}

\item{mode}{an integer value from 1 to 3 representing the category of colors}
}
\value{
a vector of colors
}
\description{
get_my_colors generates a vector of colors
}
\seealso{
\code{\link{plot_sub_image}};
}
\author{
Shijia Zhu, \email{shijia.zhu@UTsouthwestern.edu}
}
