% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_STcellchat.R
\name{get_STcellchat}
\alias{get_STcellchat}
\title{get_STcellchat}
\usage{
get_STcellchat(STIE_result, ST_expr, database = c("human", "mouse"),
  db_category = c("Secreted Signaling"), max_reps = NULL,
  nboot = 100, thresh = 0.05)
}
\arguments{
\item{STIE_result}{a list representing the result generated by the STIE() function}

\item{ST_expr}{a matrix of gene expression representing the spot level gene expression, with row being the spot and column being the gene}

\item{database}{a character value in c("human","mouse") representing the database of STRINGdb protein-protein interaction and CellChat ligand-receptor interaction}

\item{db_category}{a character value representing the category of pathways in the CellChat ligand-receptor interaction database}

\item{max_reps}{a vector of integer values representing the number of cells in each cell type. When max_reps=NULL, it takes the number of cells for each cell type from STIE_result}

\item{nboot}{an integer representing the number of bootstrapping for calculating pvalue}

\item{thresh}{a numeric value representing the pvalue threshold for the cell-cell interaction probability}
}
\value{
an object of CellChat
}
\description{
get_STcellchat generates a CellChat object from the spot level gene expression and STIE estimation
}
\references{

}
\seealso{
\code{\link{plot_STcellchat}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
