% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{getSummary}
\usage{
get_summary(STIE_result, ST_expr)
}
\arguments{
\item{STIE_result}{a list generated by the STIE() function}

\item{ST_expr}{a matrix of spot level gene expression with row representing the spot and column representing the gene}
}
\value{
A list containing the follow components:
\itemize{
 \item {rmse} a vector of numeric values representing RMSE for each spot
 \item {mse} a data frame representing the cells on spots along with the cellular morphological features, which is the same data frame cells_on_spot with the input
 \item {LL_Morp} a vector of numeric values representing RMSE for each spot
 \item {LL_Expr} a data frame representing the cells on spots along with the cellular morphological features, which is the same data frame cells_on_spot with the input
 \item {logLik} a data frame representing the cells on spots along with the cellular morphological features, which is the same data frame cells_on_spot with the input
 \item {EM_diff2} a numeric value representing difference between morphology and expression-based estimation
 \item {L2_sum} a numeric value representing MSE + lambda*EM_diff2
}
}
\description{
getSummary calculates the scores for the STIE result
}
\seealso{
\code{\link{STIE}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
