% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSTdata.R
\name{load_STdata}
\alias{load_STdata}
\title{load_STdata}
\usage{
load_STdata(sample_names, sample_dir, umi_cutoff = 0,
  is_normalized = TRUE)
}
\arguments{
\item{sample_names}{}

\item{sample_dir}{}

\item{umi_cutoff}{}

\item{is_normalized}{}
}
\description{
load_STdata
}
