% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netVisual_STcellchat.R
\name{netVisual_STcellchat}
\alias{netVisual_STcellchat}
\title{netVisual_STcellchat}
\usage{
netVisual_STcellchat(cellchat, cell_dist, cell_types, color_use = NULL)
}
\arguments{
\item{cellchat}{a CellChat object generated by the STIE::get_STcellchat() function}

\item{cell_dist}{a matrix representing the distance between cell pairs}

\item{cell_types}{a vector of character values representing the cell type, with the name being the cell_id}

\item{color_use}{a vector of character values representing the colors of cell types. The cell types are ranked based on the alphabetical order}
}
\description{
netVisual_STcellchat draws the circos plot for cellular communication with spatial informatioin taken into account
}
\references{

}
\seealso{
\code{\link{get_STcellchat}}; \code{\link{plot_STcellchat}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
