% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sub_image.R
\name{plot_cell_contour}
\alias{plot_cell_contour}
\title{plot_cell_contour}
\usage{
plot_cell_contour(contour, cell_coordinates, w, h, xoff, yoff, x_scale,
  cell_cols)
}
\arguments{
\item{contour}{a list of matrics with each item representing one cell contour}

\item{cell_coordinates}{a matix of numeric values represensting the coordinate of each cell, where columns "pixel_x" and "pixel_y" represent the x and y coordinates, respectively}

\item{w}{an integer value representing the width for cropping the image}

\item{h}{an integer value representing the height for cropping the image}

\item{xoff}{an integer value representing the offset on X-axis for cropping the image}

\item{yoff}{an integer value representing the offset on Y-axis for cropping the image}

\item{x_scale}{a numeric value from 0 to 1 for resizing the image}

\item{cell_cols}{a vector of character values representing colors to plot cell types, which ara ranked based on alphabetical order}
}
\description{
plot_cell_contour
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
