% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_STcellchat.R
\name{plot_STcellchat}
\alias{plot_STcellchat}
\title{plot_STcellchat}
\usage{
plot_STcellchat(cellchat, cell_dist, im, image_transparency = 0,
  w = NULL, h = NULL, xoff = 0, yoff = 0, x_scale = 1,
  plot_object, direction = c("outgoing", "incoming"), prob_cutoff = 0,
  color_use = NULL, lwd = 10, plot_cell = FALSE, contour = NULL)
}
\arguments{
\item{cellchat}{a CellChat object generaged by STIE::get}

\item{cell_dist}{a data frame representing the distance between cell pairs}

\item{im}{an image object generated by the magick::image_read() function}

\item{image_transparency}{a numeric value from 0 to 1 representing the transparency for plotting the image}

\item{w}{an integer value representing the width for cropping the image}

\item{h}{an integer value representing the height for cropping the image}

\item{xoff}{an integer value representing the offset on X-axis for cropping the image}

\item{yoff}{an integer value representing the offset on Y-axis for cropping the image}

\item{x_scale}{a numeric value from 0 to 1 for resizing the image}

\item{plot_object}{a character value representing the pathway or pattern of interest to plot}

\item{direction}{a character value taking "outgoing" or "incoming", representing the sending or receiving signal for the plot_object}

\item{prob_cutoff}{a numeric value representing the probability cutoff of interaction signal for plotting the plot_object}

\item{color_use}{a vector of character values representing colors to plot cell types, which ara ranked based on alphabetical order}

\item{lwd}{a numeric value representing the thickness of edge for cell-cell interaction}

\item{plot_cell}{a boolean value representing whether to plot cell contours}

\item{contour}{a list of matrics with each item representing one cell contour}
}
\description{
plot_STcellchat draws the spatially resolved cell-cell interaction onto the image
}
\seealso{
\code{\link{get_STcellchat}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
