% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sub_image.R
\name{plot_sub_image}
\alias{plot_sub_image}
\title{plot_sub_image}
\usage{
plot_sub_image(im = NULL, im_path = NULL, w = NULL, h = NULL,
  xoff = 0, yoff = 0, x_scale = 1, plot_spot = F, fct = 0.25,
  spot_coordinates = NULL, spot_types = NULL, plot_cell = T, contour,
  cell_types, axis_tick = 0, axis_col = "grey")
}
\arguments{
\item{im}{}

\item{im_path}{}

\item{w}{}

\item{h}{}

\item{xoff}{}

\item{yoff}{}

\item{x_scale}{}

\item{plot_spot}{}

\item{fct}{}

\item{spot_coordinates}{}

\item{spot_types}{}

\item{plot_cell}{}

\item{contour}{}

\item{cell_types}{}

\item{axis_tick}{}

\item{axis_col}{}
}
\description{
plot_sub_image
}
