% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{read_image}
\alias{read_image}
\title{read_image}
\usage{
read_image(image, croparea = NULL, x_scale = 1)
}
\arguments{
\item{image}{a character value indicating the path to the image}

\item{croparea}{a character value in the format of "width,height,xoff,yoff" for cropping the image}

\item{x_scale}{a numeric value from 0 to 1 for resizing the image}
}
\value{
an EBImage object of the cropped and resized image
}
\description{
read_image loads the region of interest from an image at a predefined rescaled size
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
