% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescue_missing_cells.R
\name{rescue_missing_cells}
\alias{rescue_missing_cells}
\title{rescue_missing_cells identify cell types for the missing cells falling in gap area between spots based on the learned STIE model}
\usage{
rescue_missing_cells(STIE_result, cells_on_spot_new,
  merge_STIE_result = TRUE)
}
\arguments{
\item{STIE_result}{a STIE object generated by the STIE() function}

\item{cells_on_spot_new}{a data frame by the get_cells_on_spot() function representing missing cells that are not in STIE_result}

\item{merge_STIE_result}{a boolean value indicating whether to merge the result of missing cells with STIE_result}
}
\value{
A list containing the follow components:
\itemize{
 \item {lambda} a numeric value representing the shrinkage penalty of nuclear morphology, which is the same data frame cells_on_spot with the input
 \item {mu} a vector of numeric values representing the means of morphological features
 \item {sigma} a vector of numeric values representing the standard deviations of morphological features
 \item {PE_on_spot} a matrix of cell type probabilities for each spot estimated from the spot gene expression, where the row is the spot and column is the cell type
 \item {PM_on_cell} a matrix of cell type probabilities for each cell estimated from the cellular morphological features, where the row is the cell and column is the cell type
 \item {PME_uni_cell} a matrix of cell type probabilities for each cell estimated from both spot gene expression and cellular morphological features, where the row is the cell and column is the cell type
 \item {cell_types} a vector of character values representing the cell type for each cell, where the names(cell_types) are the unique cell ids
 \item {uni_cell_types} a vector of character values representing the cell type of unqiue cells with redundant cells eliminated 
 \item {Signature} a matrix of gene expression in gene X cell type. During deconvolution, it is the same with the input. During clustering, it is the re-estimated signature from the ST data. 
 \item {cells_on_spot} a data frame representing the cells on spots along with the cellular morphological features, which is the same data frame cells_on_spot with the input
}
}
\description{
rescue_missing_cells identify cell types for the missing cells falling in gap area between spots based on the learned STIE model
}
\references{

}
\seealso{
\code{\link{get_cells_on_spot}}; \code{\link{STIE}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
