% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_ini.R
\name{signature_ini}
\alias{signature_ini}
\title{signature_ini}
\usage{
signature_ini(k, ST_expr = NULL, cells_on_spot = NULL, on = c("expr",
  "morph"), features = c("size", "shape"), pc_n = 10,
  rho_cutoff = 0.99)
}
\arguments{
\item{k}{integer value indicating the number of clusters}

\item{ST_expr}{a matrix of gene expression representing the spot level gene expression, with row being the spot and column being the gene}

\item{cells_on_spot}{a data frame representing the cells on spots along with the cellular morphological features. It will be used when on=='morph'}

\item{on}{a character value taking from c("expr",'morph'), indicating whether to use gene expression initial value at spot level or morphological value at the single-cell level}

\item{features}{a vector of character values indicating what mophorlogical features to use, when on=='morph',}

\item{pc_n}{an integer value indicating the number of PCs used for K-means on gene expression, when on=='expr' (by default pc_n=10)}

\item{rho_cutoff}{a numeric value indicating the cutoff for the similarity between transcriptomic signatures. Once greater than the cutoff, one of the signature is removed, due to redundancy.}
}
\value{
a matrix of gene expression representing the cell type transcriptomic signature, with row being the gene and column being the cell type
}
\description{
signature_ini assigns initial values of transcriptomic signature for STIE clustering
}
\examples{
Signature_ = signature_ini(k=5, ST_expr=ST_expr, on="expr")
Signature_ = signature_ini(k=5, cells_on_spot=cells_on_spot, on='morph', features=c('size','shape'))


}
