% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_image.R
\name{split_image}
\alias{split_image}
\title{split_image}
\usage{
split_image(image, split_dir, w = 3000, h = 3000, margin = 100,
  x_scale = 1, spot_coordinates = NULL, rgbscale = FALSE,
  grayscale = FALSE)
}
\arguments{
\item{image}{}

\item{split_dir}{}

\item{w}{}

\item{h}{}

\item{margin}{}

\item{x_scale}{}

\item{spot_coordinates}{}

\item{rgbscale}{}

\item{grayscale}{}
}
\description{
split_image
}
