% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STIE.R
\name{STIE}
\alias{STIE}
\title{STIE}
\usage{
STIE(ST_expr, Signature, cells_on_spot, features, lambda = 0,
  steps = 30, morphology_steps = ceiling(steps/3),
  known_signature = TRUE, known_cell_types = FALSE, min_cells = 2)
}
\arguments{
\item{ST_expr}{}

\item{Signature}{}

\item{cells_on_spot}{}

\item{features}{}

\item{lambda}{}

\item{steps}{}

\item{morphology_steps}{}

\item{known_signature}{}

\item{min_cells}{}
}
\description{
STIE
}
\examples{


}
