% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STIE_search.R
\name{STIE_search}
\alias{STIE_search}
\title{STIE_search}
\usage{
STIE_search(ST_expr, Signature, cells_on_spot, steps = 30,
  morphology_steps = ceiling(steps/3), known_signature = TRUE,
  known_cell_types = FALSE, min_cells = 2, equal_prior = TRUE,
  lambdas = c(0, 1000, 1e+06), features_of_interest = c("size",
  "shape"), criterion = c("L2sum", "rmse", "logLik"), plot = TRUE)
}
\arguments{
\item{ST_expr}{a matrix of gene expression representing the spot level gene expression, with row being the spot and column being the gene}

\item{Signature}{a matrix of gene expression representing the cell type transcriptomic signature, with row being the gene and column being the cell type}

\item{cells_on_spot}{a data frame representing the cells on spots along with the cellular morphological features, where "cell_id" represents the unique cell id and "spot" represents the uniqe spot id}

\item{steps}{an integer value representing the number of iterative steps of EM algorithm}

\item{morphology_steps}{an integer value representing the number of iterative steps of updating mophological model before updating gene expression model}

\item{known_signature}{a boolean value representing whether the transcriptomic signature is given. If TRUE, STIE performs single cell deconvolution}

\item{known_cell_types}{a boolean value representing whether the cell typing is given. If TRUE, STIE will perform cell type signature learning}

\item{min_cells}{a boolean value representing the minimum number of cells to keep for each cell type. If the cell count is smaller than min_cells, the cell type will be eliminated}

\item{equal_prior}{a boolean value representing the wheter or not assume equal prior, i.e., P(q|theta)}

\item{lambdas}{a vector of numeric value representing the shrinkage penalty of nuclear morphology}

\item{plot}{a boolean value representing whether to plot the result}

\item{features}{a vector of character values, representing the morphological features used in the STIE model}
}
\value{
A list containing the follow components:
\itemize{
 \item {lambda} a numeric value representing the shrinkage penalty of nuclear morphology, which is the same data frame cells_on_spot with the input
 \item {mu} a vector of numeric values representing the means of morphological features
 \item {sigma} a vector of numeric values representing the standard deviations of morphological features
 \item {PE_on_spot} a matrix of cell type probabilities for each spot estimated from the spot gene expression, where the row is the spot and column is the cell type
 \item {PM_on_cell} a matrix of cell type probabilities for each cell estimated from the cellular morphological features, where the row is the cell and column is the cell type
 \item {PME_uni_cell} a matrix of cell type probabilities for each cell estimated from both spot gene expression and cellular morphological features, where the row is the cell and column is the cell type
 \item {cell_types} a vector of character values representing the cell type for each cell, where the names(cell_types) are the unique cell ids
 \item {uni_cell_types} a vector of character values representing the cell type of unqiue cells with redundant cells eliminated 
 \item {Signature} a matrix of gene expression in gene X cell type. During deconvolution, it is the same with the input. During clustering, it is the re-estimated signature from the ST data. 
 \item {cells_on_spot} a data frame representing the cells on spots along with the cellular morphological features, which is the same data frame cells_on_spot with the input
}
}
\description{
STIE_search: hyperparameter and feature selection using STIE function
}
\references{

}
\seealso{
\code{\link{get_cells_on_spot}}; \code{\link{split_image}}; \code{\link{run_imageJ_plugin}}; \code{\link{merge_feature}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
