% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_morphology_parameter.R
\name{update_morphology_parameter}
\alias{update_morphology_parameter}
\title{update_morphology_parameter}
\usage{
update_morphology_parameter(PE_on_spot, PM_on_cell, cells_on_spot,
  features)
}
\arguments{
\item{PE_on_spot}{}

\item{PM_on_cell}{}

\item{cells_on_spot}{}

\item{features}{}
}
\description{
update_morphology_parameter
}
