% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BIC.R
\name{calculate_BIC}
\alias{calculate_BIC}
\title{calculate_BIC}
\usage{
calculate_BIC(STIE_result, ST_expr)
}
\arguments{
\item{STIE_result}{a list generated by the STIE() function}

\item{ST_expr}{a matrix of spot level gene expression with row representing the spot and column representing the gene}
}
\value{
A list containing the follow components:
\itemize{
 \item {rmse} a vector of numeric values representing RMSE for each spot
 \item {cells_on_spot} a data frame representing the cells on spots along with the cellular morphological features, which is the same data frame cells_on_spot with the input
}
}
\description{
calculate_BIC calculates the RMSE and BIC score for the STIE result
}
\seealso{
\code{\link{STIE}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
