% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_dist.R
\name{calculate_cell_dist}
\alias{calculate_cell_dist}
\title{calculate_cell_dist}
\usage{
calculate_cell_dist(cells_on_spot, dist_type = c("center", "boundary"),
  dist_cutoff = Inf, axis = c("Minor", "Major"))
}
\arguments{
\item{cells_on_spot}{a data frame indicating the cells on spots}

\item{dist_type}{whether the distance is between cell centers or cell boundary}

\item{dist_cutoff}{a numeric value indicating the cutoff of maximum distance between two cells}

\item{axis}{whether the distance is calculated along major or minor axis}
}
\description{
calculate_cell_dist
}
\examples{



}
\references{

}
\seealso{
\code{\link{get_cells_on_spot}};
}
\author{
Shijia Zhu, \email{shijia.zhu@UTsouthwestern.edu}
}
