% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_dist.R
\name{calculate_cell_dist}
\alias{calculate_cell_dist}
\title{calculate_cell_dist}
\usage{
calculate_cell_dist(cells_on_spot, max_dist = Inf)
}
\arguments{
\item{cells_on_spot}{a data frame indicating the cells on spots}

\item{max_dist}{a numeric value indicating the cutoff of maximum distance between two cells}
}
\description{
calculate_cell_dist
}
\examples{



}
\references{

}
\seealso{
\code{\link{get_cells_on_spot}};
}
\author{
Shijia Zhu, \email{shijia.zhu@UTsouthwestern.edu}
}
