% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sig_markers.R
\name{find_sig_markers}
\alias{find_sig_markers}
\title{find_signature_markers}
\usage{
find_sig_markers(STIE_result, ST_expr, transform = "inverse.transform",
  DEG_pthres = 1e-05)
}
\arguments{
\item{STIE_result}{}

\item{DEG_pthres}{}

\item{database}{}

\item{db_category}{}
}
\description{
find_signature_markers
}
\examples{



}
