% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_STcellchat.R
\name{get_STcellchat}
\alias{get_STcellchat}
\title{get_STcellchat}
\usage{
get_STcellchat(STIE_result, ST_expr, database = c("human", "mouse"),
  db_category = c("Secreted Signaling"), max_reps = NULL,
  nboot = 100, thresh = 0.05)
}
\arguments{
\item{STIE_result}{}

\item{ST_expr}{}

\item{database}{}

\item{db_category}{}

\item{max_reps}{}
}
\description{
get_STcellchat
}
\examples{



}
