% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSTdata.R
\name{load_STdata}
\alias{load_STdata}
\title{load_STdata}
\usage{
load_STdata(sample_names, sample_dir, umi_cutoff = 0,
  is_normalized = TRUE)
}
\arguments{
\item{sample_names}{a character value representing the sample name}

\item{sample_dir}{a character value representing the path where spaceranger generate the read count}

\item{umi_cutoff}{an integer value representing the UMI count to filter spots}

\item{is_normalized}{a boolean value indicating whether to normalize the raw read count}
}
\value{
A list containing the follow components:
\itemize{
 \item {matrix} a matrix of raw read count on the spot
 \item {normalized_matrix} a matrix of normalized read count on the spot
 \item {bcs_merge} a data frame representing the spot information
 \item {images_tibble} an image tibble object for the ST matched image
}
}
\description{
load_STdata
}
\references{

}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
