% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netVisual_STcellchat.R
\name{netVisual_STcellchat}
\alias{netVisual_STcellchat}
\title{netVisual_STcellchat}
\usage{
netVisual_STcellchat(cellchat, cell_dist, cell_types, color_use = NULL)
}
\arguments{
\item{cellchat}{}

\item{cell_dist}{}

\item{cell_types}{}

\item{color_use}{}
}
\description{
netVisual_STcellchat
}
