% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_STcellchat.R
\name{plot_STcellchat}
\alias{plot_STcellchat}
\title{plot_STcellchat}
\usage{
plot_STcellchat(cellchat, cell_dist, im, image_transparency = 0,
  w = NULL, h = NULL, xoff = 0, yoff = 0, x_scale = 1,
  plot_object, direction = c("outgoing", "incoming"), prob_cutoff = 0,
  color_use = NULL, lwd = 10, plot_cell = FALSE, contour = NULL)
}
\arguments{
\item{cellchat}{}

\item{cell_dist}{}

\item{im}{}

\item{image_transparency}{}

\item{w}{}

\item{h}{}

\item{xoff}{}

\item{yoff}{}

\item{x_scale}{}

\item{plot_object}{}

\item{direction}{}

\item{color_use}{}

\item{lwd}{}

\item{plot_cell}{}

\item{contour}{}
}
\description{
plot_STcellchat
}
\examples{


}
