% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sub_image.R
\name{plot_sub_image}
\alias{plot_sub_image}
\title{plot_sub_image}
\usage{
plot_sub_image(im = NULL, im_path = NULL, image_transparency = 0,
  w = NULL, h = NULL, xoff = 0, yoff = 0, x_scale = 1,
  plot_spot = F, spot_radius = NULL, fct = 0.25,
  spot_coordinates = NULL, plot_cell = T, contour = NULL,
  cell_types = NULL, color_use = NULL, axis_tick = 0,
  axis_col = "grey", spot_cols = "white", fill_spot = FALSE)
}
\arguments{
\item{im}{an image object generated by the magick::image_read() function}

\item{im_path}{a character value, representing the the path of the image file}

\item{image_transparency}{a numeric value ranging from 0 to 1, representing the transparency to plot the image}

\item{w}{a numeric value representing the width for image cropping}

\item{h}{a numeric value representing the height for image cropping}

\item{xoff}{a numeric value representing the x-axis offset for image cropping}

\item{yoff}{a numeric value representing the y-axis offset for image cropping}

\item{x_scale}{a numeric value from 0 to 1, representing the scalling factor for resizing image}

\item{plot_spot}{a boolean value representing whether to plot the spot}

\item{spot_radius}{a numeric value representing the spot radius}

\item{fct}{a numeric value representing the ratio between spot radius and spot center distance (for calculating spot radius)}

\item{spot_coordinates}{a data frame representing the spot coordinates, where barcode indicates the spot id}

\item{plot_cell}{a boolean value representing whether to plot the cell contour}

\item{contour}{a list of cell contours, where each item is a data frame representing the contour coordinates}

\item{cell_types}{a vector of character values, with the names being the cell ids}

\item{color_use}{a vector of character values representing the colors to plot the cell contour}

\item{axis_tick}{a numeric values representing the interval between two ticks}

\item{axis_col}{a character value representing the color of axis}

\item{spot_cols}{a vector of character values representing the colors of spots}

\item{fill_spot}{a boolean value representing whether to fill in the spot}
}
\description{
plot_sub_image draws the spatially resolved single cells with cell typing onto the histology image
}
\seealso{
\code{\link{get_my_colors}}; \code{\link{STIE}}; \code{\link{merge_feature}};
}
\author{
Shijia Zhu, \email{shijia.zhu@utsouthwestern.edu}
}
