% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_imageJ_plugin.R
\name{run_imageJ_plugin}
\alias{run_imageJ_plugin}
\title{run_imageJ_plugin}
\usage{
run_imageJ_plugin(imageJ, plugin_macro, split_image_dir,
  feature_dir = split_image_dir, pattern = "jpg$")
}
\arguments{
\item{imageJ}{}

\item{plugin_macro}{}

\item{split_image_dir}{}

\item{feature_dir}{}

\item{pattern}{}
}
\description{
run_imageJ_plugin
}
