% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_STdata.R
\name{simulate_STdata}
\alias{simulate_STdata}
\title{simulate_STdata}
\usage{
simulate_STdata(cells_coordinates, cell_types, Signature,
  spot_coordinates_ref, spot_diameter_pixel_ref = spot_radius * 2,
  spot_size_ref = 55, spot_size_sim, spot_dist_sim = spot_size_sim * 2,
  x_scale = args$x_scale, ncore = 30)
}
\arguments{
\item{cells_coordinates}{}

\item{cell_types}{a vector of character values with names representing the cell_id}

\item{Signature}{a matrix of numeric values representing the cell type transcriptomic signature with row being the gene and column being the cell type}

\item{spot_coordinates_ref}{}

\item{spot_diameter_pixel_ref}{}

\item{spot_size_ref}{}

\item{spot_size_sim}{}

\item{spot_dist_sim}{}

\item{x_scale}{}

\item{ncore}{}
}
\description{
simulate_STdata simulates the spot level ST data at a predefined spot size from the real ST data
}
\examples{


}
