% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_regression_parameter.R
\name{update_expression_regression_parameter}
\alias{update_expression_regression_parameter}
\title{update_expression_regression_parameter}
\usage{
update_expression_regression_parameter(Signature, Expr_on_spot_i,
  PE_on_spot_i, PM_on_spot_i, lambda = c(0, 10^c(0:7)), scaled = T,
  kfold = 10)
}
\arguments{
\item{Signature}{}

\item{Expr_on_spot_i}{}

\item{PE_on_spot_i}{}

\item{PM_on_spot_i}{}

\item{lambda}{}

\item{scaled}{}

\item{kfold}{}
}
\description{
update_expression_regression_parameter
}
